requirements.txt
openai>=1.40.0
pyyaml>=6.0
requests>=2.31
rich>=13.0
click>=8.1
numpy>=1.24
scipy>=1.11
pandas>=2.0
tqdm>=4.65
jsonschema>=4.17
biopython>=1.81
